Adding untested incomplete GUI version because work will stop for minimum two weeks.
The user interface appearance is editable by glade.
Much work is done by server programs, not all of which may be in this set.
This version requires a postgres installation for which the user has the
ability to create new databases.  Results printing requires pdfroff.
Writing to DVD requires VTE, because the user is prompted through a terminal
session.